app/requirements.txt
flask==2.0.2
pillow==9.5.0
pandas==2.0.2